Enable warnings and a small refactor
