+++ src/genmap-impl.h
@@ -139,6 +139,7 @@ typedef struct {
 sint get_components(sint *component, struct rsb_element *elements,
                     struct comm *c, buffer *buf, uint nelt, uint nv);
 
+void split_and_repair_partitions(genmap_handle h, struct comm *lc, int level);
 /* Matrix inverse */
 void matrix_inverse(int N, double *A);
